requirements.txt
langchain==0.1.17
langchain-core==0.1.52
langchain-community==0.0.36
langchain-openai==0.1.6
openai==1.28.1
streamlit==1.31.1
python-dotenv==1.0.1
python-docx
nltk
PyPDF2